+++ packages/quantic/force-app/main/default/lwc/quanticDateFacet/quanticDateFacet.js
@@ -3,9 +3,12 @@ import {
   registerComponentForInit,
   initializeWithHeadless,
 } from 'c/quanticHeadlessLoader';
+import {I18nUtils} from 'c/quanticUtils';
 import LOCALE from '@salesforce/i18n/locale';
 
 import clear from '@salesforce/label/c.quantic_Clear';
+import collapseFacet from '@salesforce/label/c.quantic_CollapseFacet';
+import expandFacet from '@salesforce/label/c.quantic_ExpandFacet';
 
 export default class QuanticDateFacet extends LightningElement {
   /** @type {import("coveo").DateFacetState} */
@@ -13,6 +16,9 @@ export default class QuanticDateFacet extends LightningElement {
   @track state = {
     values: [],
   };
+
+  /** @type {string} */
+  @api facetId;
   /** @type {string} */
   @api field;
   /** @type {string} */
@@ -33,12 +39,12 @@ export default class QuanticDateFacet extends LightningElement {
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
   /** @type {boolean} */
-  isCollapsed = false;
-  /** @type {string} */
-  collapseIcon = 'utility:dash';
+  isExpanded = true;
 
   labels = {
     clear,
+    collapseFacet,
+    expandFacet,
   };
 
   connectedCallback() {
@@ -59,6 +65,7 @@ export default class QuanticDateFacet extends LightningElement {
         field: this.field,
         numberOfValues: Number(this.numberOfValues),
         generateAutomaticRanges: true,
+        facetId: this.facetId ?? this.field,
       },
     });
     this.unsubscribe = this.facet.subscribe(() => this.updateState());
@@ -93,6 +100,15 @@ export default class QuanticDateFacet extends LightningElement {
     return this.state.hasActiveValues;
   }
 
+  get actionButtonIcon() {
+    return this.isExpanded ? 'utility:dash' : 'utility:add';
+  }
+  
+  get actionButtonLabel() {
+    const label = this.isExpanded ? this.labels.collapseFacet : this.labels.expandFacet;
+    return I18nUtils.format(label, this.label);
+  }
+
   /**
    * @param {CustomEvent<import("coveo").DateFacetValue>} evt
    */
@@ -105,7 +121,10 @@ export default class QuanticDateFacet extends LightningElement {
   }
 
   toggleFacetVisibility() {
-    this.collapseIcon = this.isCollapsed ? 'utility:dash' : 'utility:add';
-    this.isCollapsed = !this.isCollapsed;
+    this.isExpanded = !this.isExpanded;
+  }
+
+  preventDefault(evt) {
+    evt.preventDefault();
   }
 }